address comments

+++ kayenta-graphite/src/main/java/com/netflix/kayenta/graphite/service/GraphiteRemoteService.java
@@ -11,7 +11,7 @@ public interface GraphiteRemoteService {
     // From https://graphite.readthedocs.io/en/1.1.2/render_api.html#
     @GET("/render")
     List<GraphiteResults> rangeQuery(@Query(value = "target", encodeValue = false) String target,
-                                    @Query("from") int from,
-                                    @Query("until") int until,
+                                    @Query("from") long from,
+                                    @Query("until") long until,
                                     @Query("format") String format);
 }
